% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{calculateIons}
\alias{calculateIons}
\title{A function to calculate Y-ions}
\usage{
calculateIons(IDPdb)
}
\arguments{
\item{IDPdb}{a data.frame containing MonoisotopicPeptideMasses and corresponding MonoisotopicY1mass, and the charge state}
}
\description{
this function calculates Y-ions
}
\examples{
MonoisotopicPeptideMass <- c(1000, 2000, 3000, 4000)
MonoisotopicY1mass <- c(1203.079, 2203.079, 3203.079, 4203.079)
charge <- c(2,5,3,4)
IDPdb <- data.frame(MonoisotopicY1mass,MonoisotopicPeptideMass, charge)
Ions <- calculateIons(IDPdb)
}
\keyword{calculate}
