% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{compileData}
\alias{compileData}
\title{A function to compile data for the 'data' argument in plantGlycoMS}
\usage{
compileData(analysis = "analysis", sampleName = "results", IDPdb = IDPdb,
  MS2Data = MS2Data, Ions = Ions)
}
\arguments{
\item{analysis}{a character string with the name of the analysis, defaults to "analysis"}

\item{sampleName}{a character string with the name of the sample, defaults to "results"}

\item{IDPdb}{a data.frame containing the peptide spectrum matches, defaults to IDPdb}

\item{MS2Data}{a list of MS2 binary text files, defaults to MS2Data}

\item{Ions, }{a list of ion m/z values for each peptide spectrum match, defaults to Ions}
}
\description{
this function compiles data for the 'data' argument in plantGlycoMS
}
\examples{
compileData()
}
\keyword{compile}
