% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{defaultIons}
\alias{defaultIons}
\title{A function to define default ion values}
\usage{
defaultIons(b, y)
}
\arguments{
\item{b}{mass of a b-ion}

\item{y}{mass of a y-ion}
}
\description{
this function defines default ion values
}
\examples{
defaultIons()
}
\keyword{protViz}
