% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_PepPlus}
\alias{find_PepPlus}
\title{A function to calculate the Y0-ions}
\usage{
find_PepPlus(MonoisotopicPeptideMass, charge)
}
\arguments{
\item{MonoisotopicPeptideMass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of a Y0-ion
}
\examples{
find_Y0(1000,4)
}
\keyword{calculate}
