% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y0NH3}
\alias{find_Y0NH3}
\title{A function to calculate the neutral loss of ammonia from a Y0- ion}
\usage{
find_Y0NH3(MonoisotopicPeptideMass, charge)
}
\arguments{
\item{MonoisotopicPeptideMass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of the neutral loss of ammonia from a Y0- ion
}
\examples{
find_Y0NH3(1000,4)
}
\keyword{calculate}
