% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y1}
\alias{find_Y1}
\title{A function to calculate the Y1-ions}
\usage{
find_Y1(MonoisotopicY1mass, charge)
}
\arguments{
\item{MonoisotopicY1mass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of a Y1-ion
}
\examples{
find_Y1(1000,4)
}
\keyword{calculate}
