% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y1Ca}
\alias{find_Y1Ca}
\title{A function to calculate the neutral loss of a carbamidomethyl group from a Y-1 ion}
\usage{
find_Y1Ca(MonoisotopicY1mass, charge)
}
\arguments{
\item{MonoisotopicY1mass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of the neutral loss of a carbamidomethyl group from a Y-1 ion
}
\examples{
find_Y1Ca(1000,4)
}
\keyword{calculate}
