% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y1F}
\alias{find_Y1F}
\title{A function to calculate the Y1- plus fucose ion}
\usage{
find_Y1F(MonoisotopicY1mass, charge)
}
\arguments{
\item{MonoisotopicY1mass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of the Y1 plus fucose ion
}
\examples{
find_Y1F(1000,4)
}
\keyword{calculate}
