% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y2}
\alias{find_Y2}
\title{A function to calculate the Y2-ion}
\usage{
find_Y2(MonoisotopicY1mass, charge)
}
\arguments{
\item{MonoisotopicY1mass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of the Y2-ion
}
\examples{
find_Y2(1000,4)
}
\keyword{calculate}
