% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y3FX}
\alias{find_Y3FX}
\title{A function to calculate the Y3- plus fucose and xylose ion}
\usage{
find_Y3FX(MonoisotopicY1mass, charge)
}
\arguments{
\item{MonoisotopicY1mass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of the Y1 plus fucose and xylose ion
}
\examples{
find_Y3FX(1000,4)
}
\keyword{calculate}
