% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{find_Y3X}
\alias{find_Y3X}
\title{A function to calculate the Y3 plus xylose-ion}
\usage{
find_Y3X(MonoisotopicY1mass, charge)
}
\arguments{
\item{MonoisotopicY1mass}{numeric value}

\item{charge}{numeric value}
}
\description{
this function calculates the m/z value of the Y3 plus xylose-ion
}
\examples{
find_Y3X(1000,4)
}
\keyword{calculate}
