% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{glycoChainSaw}
\alias{glycoChainSaw}
\title{A function to modify a chainsaw in silico digest}
\usage{
glycoChainSaw(digest.N, digest.n, carbamidomethylation = TRUE,
  methionineOxidation = TRUE, glycoOnly = TRUE)
}
\arguments{
\item{digest.N}{a chainsaw in silico digest with all glycosylation sites having a capital n (N)}

\item{digest.n}{a chainsaw in silico digest with all glycosylation sites having a lower case n (n)}

\item{carbamidomethylation}{57 Da is added to the peptide mass for every cysteine present, default=TRUE}

\item{methionineOxidation}{16 Da is added to the peptide mass for every methionine present, and a new row is added to the table so that every methionine containing peptide has a value for oxidized and not oxidized, default=TRUE}

\item{glycoOnly}{return a dataframe containing only data for peptide with glycosylation sites, default=TRUE}
}
\description{
this function adds the mass for carbamidomethylation and methionine oxidation and filters for glycopeptides
}
\examples{
glycoChainSaw()
}
\keyword{digest}
