% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{glycoChange}
\alias{glycoChange}
\title{A function to change N to n for predicted glycosylation sites}
\usage{
glycoChange(fasta)
}
\arguments{
\item{fasta}{a character string containing the amino acid sequence of a protein. A fasta file can be imported with the R package 'seqinr'.}
}
\description{
this function changes N to n for predicted glycosylation sites
}
\examples{
proteinSeq <- "DLQIGFYNQSCPSAESLVQQAVAAAFANNSGIAPGLIRMHFHDCFVRGCDASVLLDSTANNTAEKDAAPNNPSLRGFEVIAAAKSAVEAACPKTVSCADILAFAARDSAALAGNITYQVPSGRRDGNVSLASEALTNIPAPTFNATQLINSFAGKNLTADEMVTLSGAHSIGVSHCFSFLNRIYNFSNTSQVDPTLSSSYADLLRTKCPSNSTRFTPITVSLDIITPTVLDNRYYTGVQLTLGLLTSDQALVTEANLSAAVKNNADNLTAWVAKFAQAIVKMGQIQVLTGTQGEIRTNCSVVNSAS"
glycoChange(fasta=proteinSeq)
}
\keyword{digest}
