% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{glycoRQ}
\alias{glycoRQ}
\title{A function to estimate relative abunance of glycoforms}
\usage{
glycoRQ(RQ, rtTable, dir, rt.min.minus, rt.min.plus)
}
\arguments{
\item{RQ}{a data.frame}

\item{rtTable}{a data.frame with retention time information}

\item{dir}{a directory with SIC data}

\item{rt.min.minus}{retention time minus}

\item{rt.min.plus}{retention time plus}
}
\description{
this function estimates relative abunance of glycoforms
}
\examples{
glycoRQ()
}
\keyword{quantitation}
\keyword{relative}
