% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{gPSMvalidatorMaxis}
\alias{gPSMvalidatorMaxis}
\title{A function to validate and annotate plant glycopeptide-spectrum matches}
\usage{
gPSMvalidatorMaxis(data, modification, modificationName, mZmarkerIons,
  minMarkerIons = 2, itol_ppm = 15, minMarkerIntensityRatio = 2,
  peakplot06 = TRUE, validate = FALSE)
}
\arguments{
\item{data}{a list compiled with compileData function}

\item{modification}{vector containing the modification mass}

\item{modificationName}{a character vector containing the name of the modification}

\item{mZmarkerIons}{a vector containing the mass values of the oxonium ions}

\item{minMarkerIons}{the minimum number of oxonium ions required for a match to be valid, default is 2}

\item{itol_ppm}{the MS ion tolerance in ppm, default is 15}

\item{minMarkerIntensityRatio}{the minimum intensity ratio of marker ions, default is 2}

\item{peakplot06}{if true, plots will be generated, default is TRUE}

\item{validate}{if true, gPSMs will be validated, if false, all will be plotted, default is FALSE}
}
\description{
this function validates and annotates plant glycopeptide-spectrum matches
}
\examples{
gPSMvalidator()
}
\keyword{validation}
