% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{peakplot02}
\alias{peakplot02}
\title{A function to plot MS data}
\usage{
peakplot02(peptideSequence, spec, FUN = defaultIons,
  fi = protViz::fragmentIon(peptideSequence, FUN = FUN)[[1]], main = NULL,
  sub = NULL, xlim = range(spec$mZ, na.rm = TRUE),
  ylim = range(spec$intensity, na.rm = TRUE), itol = 0.02,
  pattern.abc = "[abc].*", pattern.xyz = "[xyz].*", ion.axes = TRUE)
}
\arguments{
\item{peptideSequence}{a peptide sequence}

\item{spec}{a spectrum}

\item{FUN}{a function to calculate ions, default is defaultIons}

\item{fi}{a fragment ion function, default is protViz::fragmentIon(peptideSequence, FUN = FUN)[[1]]}

\item{main}{the main title of a plot, default is NULL}

\item{sub}{the subtitle of a plot, default is NULL}

\item{xlim}{the limits of the x axis, default is range(spec$mZ, na.rm = TRUE)}

\item{itol}{the fragment ion mass tolerance in delta mz, default is 0.02}

\item{pattern.abc}{the abc pattern, default is "[abc].*"}

\item{pattern.xyz}{the xyz pattern, default is "[xyz].*", ion.axes = TRUE)}

\item{the}{limits of the y axis, default is ylim = range(spec$intensity, na.rm = TRUE)}
}
\description{
this function plots MS data
}
\examples{
peakplot02()
}
\keyword{protViz}
