% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{PreferredDeltaMasses}
\alias{PreferredDeltaMasses}
\title{PreferredDeltaMasses}
\usage{
PreferredDeltaMasses(input)
}
\arguments{
\item{input}{A vector of glycan masses (from the pGlycoFilter function)}
}
\value{
A vector of glycan masses with "N!{P}[ST] * " pasted in front of the glycan masses
}
\description{
this function is used to make a vector of glycan masses with "N!{P}[ST] * " pasted in front of the glycan masses which can be pasted into a MyriMatch configuration file.
}
