% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{psm06}
\alias{psm06}
\title{A function to match peptides to spectra}
\usage{
psm06(sequence, spec, FUN = defaultIon, plot = FALSE,
  fi = protViz::fragmentIon(sequence, FUN = FUN)[[1]],
  fragmentIonError = 0.06)
}
\arguments{
\item{sequence}{the peptide sequence}

\item{spec}{a spectrum}

\item{FUN}{a function for calculating ions, default is defaultIon}

\item{plot}{should the psm be plotted, default is FALSE}

\item{fi}{how are fragment ions calculated, default is protViz::fragmentIon(sequence, FUN = FUN)[[1]]}

\item{fragmentIonError}{the mass error for fragment ions in mz, default is 0.02}
}
\description{
this function plots MS data
}
\examples{
psm06()
}
\keyword{protViz}
