% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.GlycoMod.maxis}
\alias{Read.GlycoMod.maxis}
\title{A function to import GlycoMod data from maxis results}
\usage{
Read.GlycoMod.maxis(input, ChainSaw, spectrum.table = spectrum.table,
  dir = "MS2Data")
}
\arguments{
\item{input}{glycopeptide identifications from GlycoMod}

\item{ChainSaw}{an in silico digest}

\item{spectrum.table}{a summary table of MS data}

\item{dir}{a directory with MS2 binary data}
}
\description{
this function is for importing GlycoMod data
}
\examples{
Read.GlycoMod.maxis()
}
\keyword{import}
