% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.IDPdb}
\alias{Read.IDPdb}
\title{A function to import IDPicker results}
\usage{
Read.IDPdb(IDPdb, ChainSaw, dir)
}
\arguments{
\item{IDPdb}{a data.frame containing peptide spectrum matches}

\item{ChainSaw}{a data.frame of insilico digest results (see glycoChainSaw)}

\item{dir}{a directory containing MS2 binary data}
}
\description{
This function imports IDPicker results and calculates relevant values.
}
\examples{
Read.IDPdb()
}
\keyword{data}
\keyword{import}
