% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.RQ}
\alias{Read.RQ}
\title{A function to import RQ data}
\usage{
Read.RQ(input, dir = "RQ/")
}
\arguments{
\item{input}{the data}

\item{dir}{the directory with the SIC data}
}
\description{
this function is used to import RQ data and associate it with SIC data
}
\examples{
Read.RQ()
}
\keyword{import}
\keyword{quantitation}
\keyword{relative}
