% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.RQinput}
\alias{Read.RQinput}
\title{A function to make the RQ input table}
\usage{
Read.RQinput(gPSMs.ALL)
}
\arguments{
\item{gPSMs.ALL}{a data.frame with all gPSM data}
}
\description{
this function makes the RQ input table
}
\examples{
Read.RQinput()
}
\keyword{import}
