% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.SICData}
\alias{Read.SICData}
\title{A function to import SIC data}
\usage{
Read.SICData(dir = "quantitation_Chym1_ELUTE", Asn)
}
\arguments{
\item{dir}{the directory with SIC data}

\item{Asn}{later it will be a way to designate different glycoform series}
}
\description{
this function is used to import with SIC data generated with msaccess (ProteoWizard)
}
\examples{
Read.SICData()
}
\keyword{import}
\keyword{quantitation}
\keyword{relative}
