% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{retentionTimeTable}
\alias{retentionTimeTable}
\title{A function to make a retention time table}
\usage{
retentionTimeTable(dat)
}
\arguments{
\item{dat}{a data.frame with gPSM data}
}
\description{
this function filters gPSMs by retention time
}
\examples{
retentionTimeTable()
}
\keyword{table}
