% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{rt.restrict}
\alias{rt.restrict}
\title{A function to filter gPSMs by retention time}
\usage{
rt.restrict(IDPdb, rtTable, rt.min.minus, rt.min.plus)
}
\arguments{
\item{IDPdb}{a table with peptide spectrum match information generated with IDPicker}

\item{rtTable}{a table with information about identified glycopeptides and their retention times}

\item{rt.min.minus}{amount of minutes to subtract from the retention time minimum}

\item{rt.min.plus}{amount of minutes to add to the retention time maximum}
}
\description{
this function filters gPSMs by retention time
}
\examples{
rt.restrict()
}
\keyword{retentionTime}
