% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{spectrumTable}
\alias{spectrumTable}
\title{A function to calculate monoisotopic precursor mass}
\usage{
spectrumTable(input)
}
\arguments{
\item{input}{a csv file with one column named precursorMZ}
}
\description{
this function calculates the monoisotopic precursor mass from then precursor mz
}
\examples{
spectrumTable()
}
\keyword{calculate}
