% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{PreferredDeltaMasses}
\alias{PreferredDeltaMasses}
\title{A function to add the glycosylation motif for a MyriMatch database search config}
\usage{
PreferredDeltaMasses(input)
}
\arguments{
\item{input}{a vector of dynamic modification masses}
}
\description{
this function adds N!{P}[ST] * in front of a mass value
}
\keyword{data}
\keyword{import}
