% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.RQ}
\alias{Read.RQ}
\title{A function to import RQ data}
\usage{
Read.RQ(input = "Output/quantitation_Chym1_ELUTE_211.csv",
  dir = "quantitation_Chym1_ELUTE")
}
\arguments{
\item{input}{a csv file containing exact precursor mass values}

\item{dir}{a directory containing SIC data}
}
\description{
this function is used to import RQ data and associate it with SIC data
}
\examples{
Read.RQ()
}
\keyword{import}
\keyword{quantitation}
\keyword{relative}
