% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantGlycoMSFunctions.R
\name{Read.SICData}
\alias{Read.SICData}
\alias{Read.SICData}
\title{A function to import SIC data}
\usage{
Read.SICData(dir = "quantitation_Chym1_ELUTE", Asn = 211)

Read.SICData(dir = "quantitation_Chym1_ELUTE", Asn = 211)
}
\arguments{
\item{dir}{a directory containing SIC data}

\item{Asn}{PLACEHOLDER}

\item{dir}{the directory with SIC data}
}
\description{
this function is used to import with SIC data generated with msaccess (ProteoWizard)

this function is used to import with SIC data generated with msaccess (ProteoWizard)
}
\examples{
Read.SICData()
Read.SICData()
}
\keyword{import}
\keyword{quantitation}
\keyword{relative}
